BT 1.7 compat, refactor, bugfix NRE when newly spawned vehicle attacks
