+++ ColorFloaties.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Linq;
+using BattleTech.Network;
 using BattleTech.UI;
 using Harmony;
 using TMPro;
@@ -12,48 +14,55 @@ namespace PanicSystem
     {
         internal static void Colorize(CombatHUDFloatieStack __instance)
         {
-            // default outline width is zero.  have to plug a dummy value into outline colour though...
-            void SetStyle(CombatHUDStatusStackItem floatie, Color32 inner, Color32 outline, float width = 0f)
+            try
             {
-                var tmp = Traverse.Create(floatie).Field("Text").GetValue<TextMeshProUGUI>();
-                Traverse.Create(tmp).Method("SetFaceColor", inner).GetValue();
-                Traverse.Create(tmp).Method("SetOutlineColor", outline).GetValue();
-                Traverse.Create(tmp).Method("SetOutlineThickness", width).GetValue();
-            }
+                // default outline width is zero.  have to plug a dummy value into outline colour though...
+                void SetStyle(CombatHUDStatusStackItem floatie, Color32 inner, Color32 outline, float width = 0f)
+                {
+                    var tmp = Traverse.Create(floatie).Field("Text").GetValue<TextMeshProUGUI>();
+                    Traverse.Create(tmp).Method("SetFaceColor", inner).GetValue();
+                    Traverse.Create(tmp).Method("SetOutlineColor", outline).GetValue();
+                    Traverse.Create(tmp).Method("SetOutlineThickness", width).GetValue();
+                }
 
-            var floaties = __instance.gameObject.GetComponentsInChildren<CombatHUDStatusStackItem>();
-            foreach (var floatie in floaties)
+                var floaties = __instance.gameObject.GetComponentsInChildren<CombatHUDStatusStackItem>();
+                foreach (var floatie in floaties)
+                {
+                    var text = Traverse.Create(floatie).Field("Text").GetValue<TextMeshProUGUI>().text;
+                    if (PanicSystem.ejectPhraseList.Any(x => x == text))
+                    {
+                        SetStyle(floatie, Color.white, Color.red, 0.1f);
+                    }
+                    else if (text.Contains("IMPROVED TO"))
+                    {
+                        SetStyle(floatie, Color.white, Color.blue, 0.1f);
+                    }
+                    else if (text == "PANIC LEVEL CRITICAL!")
+                    {
+                        SetStyle(floatie, Color.red, Color.yellow, 0.1f);
+                    }
+                    else if (text == "Panicked")
+                    {
+                        SetStyle(floatie, Color.red / 1.25f, Color.black);
+                    }
+                    else if (text == "Stressed")
+                    {
+                        SetStyle(floatie, Color.yellow / 1.25f, Color.red);
+                    }
+                    else if (text == "Unsettled")
+                    {
+                        SetStyle(floatie, Color.gray / 1.35f, Color.black);
+                    }
+                    // need to do this because the game leaves some ... reused objects maybe??
+                    else
+                    {
+                        SetStyle(floatie, Color.white, Color.black);
+                    }
+                }
+            }
+            catch (Exception ex)
             {
-                var text = Traverse.Create(floatie).Field("Text").GetValue<TextMeshProUGUI>().text;
-                if (PanicSystem.ejectPhraseList.Any(x => x == text))
-                {
-                    SetStyle(floatie, Color.white, Color.red, 0.1f);
-                }
-                else if (text.Contains("IMPROVED TO"))
-                {
-                    SetStyle(floatie, Color.white, Color.blue, 0.1f);
-                }
-                else if (text == "PANIC LEVEL CRITICAL!")
-                {
-                    SetStyle(floatie, Color.red, Color.yellow, 0.1f);
-                }
-                else if (text == "Panicked")
-                {
-                    SetStyle(floatie, Color.red / 1.25f, Color.black);
-                }
-                else if (text == "Stressed")
-                {
-                    SetStyle(floatie, Color.yellow / 1.25f, Color.red);
-                }
-                else if (text == "Unsettled")
-                {
-                    SetStyle(floatie, Color.gray / 1.35f, Color.black);
-                }
-                // need to do this because the game leaves some ... reused objects maybe??
-                else
-                {
-                    SetStyle(floatie, Color.white, Color.black);
-                }
+                Logger.LogError(ex);
             }
         }
     }
