+++ PanicSystem.cs
@@ -23,7 +23,7 @@ namespace PanicSystem
         internal static string storageJsonPath; //store our meta trackers here
         internal static string modDirectory;
         internal static List<string> ejectPhraseList = new List<string>();
-        internal static HarmonyInstance harmony;
+        private static HarmonyInstance harmony;
 
         public static void Init(string modDir, string settings)
         {
@@ -66,6 +66,11 @@ namespace PanicSystem
             }
 
             Patches.ManualPatching();
+        }
+
+        internal static void Log(object input)
+        {
+            //FileLog.Log($"[PanicSystem] {input}");
         }
 
         /// <summary>
@@ -144,32 +149,20 @@ namespace PanicSystem
         /// </summary>
         private static float GetPanicModifier(PanicStatus pilotStatus)
         {
+            // ReSharper disable once SwitchStatementMissingSomeCases
             switch (pilotStatus)
             {
-                case PanicStatus.Unsettled:
-                {
-                    return modSettings.UnsettledPanicModifier;
-                }
-
-                case PanicStatus.Stressed:
-                {
-                    return modSettings.StressedPanicModifier;
-                }
-
-                case PanicStatus.Panicked:
-                {
-                    return modSettings.PanickedPanicModifier;
-                }
-
-                default:
-                    return 1f;
+                case PanicStatus.Unsettled: return modSettings.UnsettledPanicModifier;
+                case PanicStatus.Stressed: return modSettings.StressedPanicModifier;
+                case PanicStatus.Panicked: return modSettings.PanickedPanicModifier;
+                default: return 1f;
             }
         }
 
         /// <summary>
         /// true is a successful saving throw
         /// </summary>
-        public static float GetSavingThrow(Mech defender, Mech attacker)
+        public static float GetSavingThrow(Mech defender, AbstractActor attacker)
         {
             var pilot = defender.GetPilot();
             var weapons = defender.Weapons;
@@ -178,121 +171,129 @@ namespace PanicSystem
             float totalMultiplier = 0;
 
             DrawHeader();
-
             LogDebug($"{$"Mech health {MechHealth(defender):#.##}%",-20} | {"",10} |");
-
-            if (attacker != null && modSettings.QuirksEnabled && attacker.MechDef.Chassis.ChassisTags.Contains("mech_quirk_distracting"))
+            try
             {
-                totalMultiplier += modSettings.DistractingModifier;
-                LogDebug($"{"Distracting mech",-20} | {modSettings.DistractingModifier,10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (modSettings.HeatDamageModifier > 0)
-            {
-                totalMultiplier += modSettings.HeatDamageModifier * Patches.heatDamage;
-                LogDebug($"{$"Heat damage {Patches.heatDamage}",-20} | {modSettings.HeatDamageModifier * Patches.heatDamage,10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (PercentPilot(pilot) < 1)
-            {
-                totalMultiplier += modSettings.PilotHealthMaxModifier * PercentPilot(pilot);
-                LogDebug($"{"Pilot injuries",-20} | {modSettings.PilotHealthMaxModifier * PercentPilot(pilot),10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (defender.IsUnsteady)
-            {
-                totalMultiplier += modSettings.UnsteadyModifier;
-                LogDebug($"{"Unsteady",-20} | {modSettings.UnsteadyModifier,10} | {totalMultiplier,10:#.###}");
-            }
-
-            if (defender.IsFlaggedForKnockdown)
-            {
-                totalMultiplier += modSettings.UnsteadyModifier;
-                LogDebug($"{"Knockdown",-20} | {modSettings.UnsteadyModifier,10} | {totalMultiplier,10:#.###}");
-            }
-
-            if (modSettings.OverheatedModifier > 0 && defender.OverheatLevel < defender.CurrentHeat)
-            {
-                totalMultiplier += modSettings.OverheatedModifier;
-                LogDebug($"{"Heat",-20} | {modSettings.OverheatedModifier,10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (modSettings.ShutdownModifier > 0 && defender.IsShutDown)
-            {
-                totalMultiplier += modSettings.ShutdownModifier;
-                LogDebug($"{"Shutdown",-20} | {modSettings.ShutdownModifier,10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (PercentHead(defender) < 1)
-            {
-                totalMultiplier += modSettings.HeadMaxModifier * PercentHead(defender);
-                LogDebug($"{"Head",-20} | {modSettings.HeadMaxModifier * PercentHead(defender),10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (PercentCenterTorso(defender) < 1)
-            {
-                totalMultiplier += modSettings.CenterTorsoMaxModifier * (1 - PercentCenterTorso(defender));
-                LogDebug($"{"CT",-20} | {modSettings.CenterTorsoMaxModifier * (1 - PercentCenterTorso(defender)),10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (PercentLeftTorso(defender) < 1)
-            {
-                totalMultiplier += modSettings.SideTorsoMaxModifier * (1 - PercentLeftTorso(defender));
-                LogDebug($"{"LT",-20} | {modSettings.SideTorsoMaxModifier * (1 - PercentLeftTorso(defender)),10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (PercentRightTorso(defender) < 1)
-            {
-                totalMultiplier += modSettings.SideTorsoMaxModifier * (1 - PercentRightTorso(defender));
-                LogDebug($"{"RT",-20} | {modSettings.SideTorsoMaxModifier * (1 - PercentRightTorso(defender)),10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (PercentLeftLeg(defender) < 1)
-            {
-                totalMultiplier += modSettings.LeggedMaxModifier * (1 - PercentLeftLeg(defender));
-                LogDebug($"{"LL",-20} | {modSettings.LeggedMaxModifier * (1 - PercentLeftLeg(defender)),10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            if (PercentRightLeg(defender) < 1)
-            {
-                totalMultiplier += modSettings.LeggedMaxModifier * (1 - PercentRightLeg(defender));
-                LogDebug($"{"RL",-20} | {modSettings.LeggedMaxModifier * (1 - PercentRightLeg(defender)),10:#.###} | {totalMultiplier,10:#.###}");
-            }
-
-            // weaponless
-            if (weapons.TrueForAll(w => w.DamageLevel != ComponentDamageLevel.Functional || !w.HasAmmo)) // only fully unusable
-            {
-                if (Random.Range(1, 5) == 1) // 20% chance of appearing
+                if (modSettings.QuirksEnabled && attacker is Mech mech &&
+                    mech.MechDef.Chassis.ChassisTags.Contains("mech_quirk_distracting"))
                 {
-                    SaySpamFloatie(defender, "NO WEAPONS!");
+                    totalMultiplier += modSettings.DistractingModifier;
+                    LogDebug($"{"Distracting mech",-20} | {modSettings.DistractingModifier,10:#.###} | {totalMultiplier,10:#.###}");
                 }
 
-                totalMultiplier += modSettings.WeaponlessModifier;
-                LogDebug($"{"Weaponless",-20} | {modSettings.WeaponlessModifier,10} | {totalMultiplier,10:#.###}");
-            }
-
-            // alone
-            if (defender.Combat.GetAllAlliesOf(defender).TrueForAll(m => m.IsDead || m == defender as AbstractActor))
-            {
-                if (Random.Range(1, 5) == 0) // 20% chance of appearing
+                if (modSettings.HeatDamageModifier > 0)
                 {
-                    SaySpamFloatie(defender, "NO ALLIES!");
+                    totalMultiplier += modSettings.HeatDamageModifier * Patches.heatDamage;
+                    LogDebug($"{$"Heat damage {Patches.heatDamage}",-20} | {modSettings.HeatDamageModifier * Patches.heatDamage,10:#.###} | {totalMultiplier,10:#.###}");
                 }
 
-                totalMultiplier += modSettings.AloneModifier;
-                LogDebug($"{"Alone",-20} | {modSettings.AloneModifier,10} | {totalMultiplier,10:#.###}");
+                if (PercentPilot(pilot) < 1)
+                {
+                    totalMultiplier += modSettings.PilotHealthMaxModifier * PercentPilot(pilot);
+                    LogDebug($"{"Pilot injuries",-20} | {modSettings.PilotHealthMaxModifier * PercentPilot(pilot),10:#.###} | {totalMultiplier,10:#.###}");
+                }
+
+                if (defender.IsUnsteady)
+                {
+                    totalMultiplier += modSettings.UnsteadyModifier;
+                    LogDebug($"{"Unsteady",-20} | {modSettings.UnsteadyModifier,10} | {totalMultiplier,10:#.###}");
+                }
+
+                if (defender.IsFlaggedForKnockdown)
+                {
+                    totalMultiplier += modSettings.UnsteadyModifier;
+                    LogDebug($"{"Knockdown",-20} | {modSettings.UnsteadyModifier,10} | {totalMultiplier,10:#.###}");
+                }
+
+                if (modSettings.OverheatedModifier > 0 && defender.OverheatLevel < defender.CurrentHeat)
+                {
+                    totalMultiplier += modSettings.OverheatedModifier;
+                    LogDebug($"{"Heat",-20} | {modSettings.OverheatedModifier,10:#.###} | {totalMultiplier,10:#.###}");
+                }
+
+                if (modSettings.ShutdownModifier > 0 && defender.IsShutDown)
+                {
+                    totalMultiplier += modSettings.ShutdownModifier;
+                    LogDebug($"{"Shutdown",-20} | {modSettings.ShutdownModifier,10:#.###} | {totalMultiplier,10:#.###}");
+                }
+
+                if (PercentHead(defender) < 1)
+                {
+                    totalMultiplier += modSettings.HeadMaxModifier * PercentHead(defender);
+                    LogDebug($"{"Head",-20} | {modSettings.HeadMaxModifier * PercentHead(defender),10:#.###} | {totalMultiplier,10:#.###}");
+                }
+
+                if (PercentCenterTorso(defender) < 1)
+                {
+                    totalMultiplier += modSettings.CenterTorsoMaxModifier * (1 - PercentCenterTorso(defender));
+                    LogDebug($"{"CT",-20} | {modSettings.CenterTorsoMaxModifier * (1 - PercentCenterTorso(defender)),10:#.###} | {totalMultiplier,10:#.###}");
+                }
+
+                if (PercentLeftTorso(defender) < 1)
+                {
+                    totalMultiplier += modSettings.SideTorsoMaxModifier * (1 - PercentLeftTorso(defender));
+                    LogDebug($"{"LT",-20} | {modSettings.SideTorsoMaxModifier * (1 - PercentLeftTorso(defender)),10:#.###} | {totalMultiplier,10:#.###}");
+                }
+
+                if (PercentRightTorso(defender) < 1)
+                {
+                    totalMultiplier += modSettings.SideTorsoMaxModifier * (1 - PercentRightTorso(defender));
+                    LogDebug($"{"RT",-20} | {modSettings.SideTorsoMaxModifier * (1 - PercentRightTorso(defender)),10:#.###} | {totalMultiplier,10:#.###}");
+                }
+
+                if (PercentLeftLeg(defender) < 1)
+                {
+                    totalMultiplier += modSettings.LeggedMaxModifier * (1 - PercentLeftLeg(defender));
+                    LogDebug($"{"LL",-20} | {modSettings.LeggedMaxModifier * (1 - PercentLeftLeg(defender)),10:#.###} | {totalMultiplier,10:#.###}");
+                }
+
+                if (PercentRightLeg(defender) < 1)
+                {
+                    totalMultiplier += modSettings.LeggedMaxModifier * (1 - PercentRightLeg(defender));
+                    LogDebug($"{"RL",-20} | {modSettings.LeggedMaxModifier * (1 - PercentRightLeg(defender)),10:#.###} | {totalMultiplier,10:#.###}");
+                }
+
+                // weaponless
+                if (weapons.TrueForAll(w => w.DamageLevel != ComponentDamageLevel.Functional || !w.HasAmmo)) // only fully unusable
+                {
+                    if (Random.Range(1, 5) == 1) // 20% chance of appearing
+                    {
+                        SaySpamFloatie(defender, "NO WEAPONS!");
+                    }
+
+                    totalMultiplier += modSettings.WeaponlessModifier;
+                    LogDebug($"{"Weaponless",-20} | {modSettings.WeaponlessModifier,10} | {totalMultiplier,10:#.###}");
+                }
+
+                // alone
+                if (defender.Combat.GetAllAlliesOf(defender).TrueForAll(m => m.IsDead || m == defender as AbstractActor))
+                {
+                    if (Random.Range(1, 5) == 0) // 20% chance of appearing
+                    {
+                        SaySpamFloatie(defender, "NO ALLIES!");
+                    }
+
+                    totalMultiplier += modSettings.AloneModifier;
+                    LogDebug($"{"Alone",-20} | {modSettings.AloneModifier,10} | {totalMultiplier,10:#.###}");
+                }
+
+                totalMultiplier -= gutsAndTacticsSum;
+                LogDebug($"{"Guts and Tactics",-20} | {$"-{gutsAndTacticsSum}",10} | {totalMultiplier,10:#.###}");
+
+                var resolveModifier = modSettings.ResolveMaxModifier *
+                                      (defender.Combat.LocalPlayerTeam.Morale - modSettings.MedianResolve) /
+                                      modSettings.MedianResolve;
+                totalMultiplier -= resolveModifier;
+                LogDebug($"{$"Resolve {defender.Combat.LocalPlayerTeam.Morale}",-20} | {resolveModifier * -1,10:#.###} | {totalMultiplier,10:#.###}");
+
+                return totalMultiplier;
             }
-
-            totalMultiplier -= gutsAndTacticsSum;
-            LogDebug($"{"Guts and Tactics",-20} | {$"-{gutsAndTacticsSum}",10} | {totalMultiplier,10:#.###}");
-
-            var resolveModifier = modSettings.ResolveMaxModifier *
-                                  (defender.Combat.LocalPlayerTeam.Morale - modSettings.MedianResolve) /
-                                  modSettings.MedianResolve;
-            totalMultiplier -= resolveModifier;
-            LogDebug($"{$"Resolve {defender.Combat.LocalPlayerTeam.Morale}",-20} | {resolveModifier * -1,10:#.###} | {totalMultiplier,10:#.###}");
-
-            return totalMultiplier;
+            catch (Exception ex)
+            {
+                // BOMB
+                LogDebug(ex);
+                return -1f;
+            }
         }
 
         private static void DrawHeader()
@@ -307,7 +308,6 @@ namespace PanicSystem
             LogDebug("Panic save failure requires eject save");
 
             DrawHeader();
-
             if (modSettings.QuirksEnabled && mech.pilot.pilotDef.PilotTags.Contains("pilot_dependable"))
             {
                 savingThrow -= modSettings.DependableModifier;
@@ -317,16 +317,12 @@ namespace PanicSystem
             // calculate result
             savingThrow = Math.Max(0f, savingThrow - modSettings.BaseEjectionResist);
             LogDebug($"{"Base ejection resist",-20} | {modSettings.BaseEjectionResist,10} | {savingThrow,10:#.###}");
-
             savingThrow = (float) Math.Round(savingThrow);
             LogDebug($"{"Eject multiplier",-20} | {modSettings.EjectChanceMultiplier,10} | {savingThrow,10:#.###}");
             var roll = Random.Range(1, 100);
-
             LogDebug(new string('-', 46));
             LogDebug($"{"Saving throw",-20} | {savingThrow,-5:###}{roll,5} | {"Roll",10}");
             LogDebug(new string('-', 46));
-
-
             if (savingThrow <= 0)
             {
                 LogDebug("Negative saving throw; skipping");
@@ -336,7 +332,6 @@ namespace PanicSystem
 
             // cap the saving throw by the setting
             savingThrow = (int) Math.Min(savingThrow, modSettings.MaxEjectChance);
-
             SaySpamFloatie(mech, $"SAVE: {savingThrow}  ROLL: {roll}!");
             if (roll >= savingThrow)
             {
@@ -344,7 +339,7 @@ namespace PanicSystem
                 SaySpamFloatie(mech, $"EJECT SAVE! HEALTH: {MechHealth(mech):#.#}%");
                 return true;
             }
-            
+
             if (modSettings.QuirksEnabled && mech.MechDef.Chassis.ChassisTags.Contains("mech_quirk_noeject"))
             {
                 LogDebug("This mech can't eject (quirk)");
@@ -379,7 +374,6 @@ namespace PanicSystem
             var index = GetPilotIndex(defender);
             savingThrow *= GetPanicModifier(trackedPilots[GetPilotIndex(defender)].panicStatus);
             LogDebug($"{"Panic multiplier",-20} | {GetPanicModifier(trackedPilots[GetPilotIndex(defender)].panicStatus),10} | {savingThrow,10:#.###}");
-
             savingThrow = (float) Math.Max(0f, Math.Round(savingThrow));
             if (!(savingThrow >= 1))
             {
@@ -388,11 +382,9 @@ namespace PanicSystem
             }
 
             var roll = Random.Range(1, 100);
-
             LogDebug(new string('-', 46));
             LogDebug($"{"Saving throw",-20} | {savingThrow,-5:###}{roll,5} | {"Roll",10}");
             LogDebug(new string('-', 46));
-
             SaySpamFloatie(defender, $"{$"SAVE:{savingThrow}",-6} {$"ROLL {roll}!",3}");
 
             // lower panic level
@@ -432,6 +424,7 @@ namespace PanicSystem
             LogDebug("Failed panic save");
             SaySpamFloatie(defender, "SAVE FAIL!");
             ApplyPanicDebuff(defender);
+
             SayStatusFloatie(defender, false);
 
             // check for crit
@@ -456,25 +449,26 @@ namespace PanicSystem
             return false;
         }
 
-        // method is called despite the setting, so it can be controlled in one place
+// method is called despite the setting, so it can be controlled in one place
         private static void SaySpamFloatie(Mech mech, string message)
         {
             if (!modSettings.FloatieSpam) return;
-
             mech.Combat.MessageCenter.PublishMessage(new AddSequenceToStackMessage(
                 new ShowActorInfoSequence(mech, message, FloatieMessage.MessageNature.Neutral, false)));
         }
 
-        // bool controls whether to display as buff or debuff
+// bool controls whether to display as buff or debuff
         private static void SayStatusFloatie(Mech mech, bool buff)
         {
             var index = GetPilotIndex(mech);
+
             var floatieString = $"{trackedPilots[index].panicStatus.ToString()}";
             if (buff)
             {
                 mech.Combat.MessageCenter.PublishMessage(new AddSequenceToStackMessage(
                     new ShowActorInfoSequence(mech, floatieString, FloatieMessage.MessageNature.Inspiration, true)));
             }
+
             else
             {
                 mech.Combat.MessageCenter.PublishMessage(new AddSequenceToStackMessage(
@@ -491,7 +485,6 @@ namespace PanicSystem
         public static bool ShouldPanic(Mech mech, AttackDirector.AttackSequence attackSequence)
         {
             if (!CanPanic(mech, attackSequence)) return false;
-
             return SufficientDamageWasDone(attackSequence);
         }
 
@@ -502,7 +495,6 @@ namespace PanicSystem
 
             // can't do stuff with vehicles and buildings
             if (!(__instance.directorSequences[0].chosenTarget is Mech)) return true;
-
             return __instance.directorSequences[0].chosenTarget?.GUID == null;
         }
 
@@ -516,7 +508,6 @@ namespace PanicSystem
             if (attackSequence == null) return false;
 
             var id = attackSequence.chosenTarget.GUID;
-
             if (!attackSequence.GetAttackDidDamage(id))
             {
                 LogDebug("No damage");
@@ -525,11 +516,9 @@ namespace PanicSystem
 
             var previousArmor = Patches.mechArmorBeforeAttack;
             var previousStructure = Patches.mechStructureBeforeAttack;
-
             LogDebug($"Damage >>> A: {attackSequence.GetArmorDamageDealt(id):#.###}" +
                      $" S: {attackSequence.GetStructureDamageDealt(id):#.###}" +
                      $" ({(attackSequence.GetArmorDamageDealt(id) + attackSequence.GetStructureDamageDealt(id)) / (previousArmor + previousStructure) * 100:#.##}%)  H: {Patches.heatDamage}");
-
             if (attackSequence.GetStructureDamageDealt(id) >= modSettings.MinimumStructureDamageRequired)
             {
                 LogDebug("Structure damage requires panic save");
@@ -544,7 +533,6 @@ namespace PanicSystem
                 LogDebug($"Took {Patches.heatDamage} heat");
             }
 
-            LogDebug($"attackSequence.GetArmorDamageDealt(id) {attackSequence.GetArmorDamageDealt(id)}\nattackSequence.GetStructureDamageDealt(id) {attackSequence.GetStructureDamageDealt(id)}\nPatches.heatDamage * modSettings.HeatDamageModifier {Patches.heatDamage * modSettings.HeatDamageModifier}");
             if (attackSequence.GetArmorDamageDealt(id) + attackSequence.GetStructureDamageDealt(id) + Patches.heatDamage * modSettings.HeatDamageModifier /
                 (previousArmor + previousStructure) +
                 100 <= modSettings.MinimumDamagePercentageRequired)
@@ -555,63 +543,6 @@ namespace PanicSystem
 
             LogDebug("Total damage requires a panic save");
             return true;
-        }
-
-        public class Settings
-        {
-            public bool Debug = false;
-            public bool EnableEjectPhrases;
-            public bool FloatieSpam = false;
-            public float EjectPhraseChance = 100;
-            public bool ColorizeFloaties = true;
-
-            // panic
-            public bool PlayersCanPanic = true;
-            public bool EnemiesCanPanic = true;
-            public float MinimumDamagePercentageRequired = 10;
-            public float MinimumStructureDamageRequired = 5;
-            public bool OneChangePerTurn = false;
-            public bool LosingLimbAlwaysPanics = false;
-            public float UnsteadyModifier = 10;
-            public float PilotHealthMaxModifier = 15;
-            public float HeadMaxModifier = 15;
-            public float CenterTorsoMaxModifier = 45;
-            public float SideTorsoMaxModifier = 20;
-            public float LeggedMaxModifier = 10;
-            public float WeaponlessModifier = 10;
-            public float AloneModifier = 10;
-            public float UnsettledAimModifier = 1;
-            public float StressedAimModifier = 1;
-            public float StressedToHitModifier = -1;
-            public float PanickedAimModifier = 2;
-            public float PanickedToHitModifier = -2;
-            public float MedianResolve = 50;
-            public float ResolveMaxModifier = 10;
-            public float DistractingModifier = 0;
-            public float OverheatedModifier = 0;
-            public float ShutdownModifier = 0;
-            public float HeatDamageModifier = 10;
-
-            //deprecated public float MechHealthAlone = 50;
-            public float MechHealthForCrit = 0.9f;
-            public float CritOver = 70;
-            public float UnsettledPanicModifier = 1f;
-            public float StressedPanicModifier = 0.66f;
-            public float PanickedPanicModifier = 0.33f;
-
-            // Quirks
-            public bool QuirksEnabled = true;
-            public float BraveModifier = 5;
-            public float DependableModifier = 5;
-
-            // ejection
-            public float MaxEjectChance = 50;
-            public float EjectChanceMultiplier = 0.75f;
-
-            // deprecated public bool ConsiderEjectingWhenAlone = false;
-            public float BaseEjectionResist = 50;
-            public float GutsEjectionResistPerPoint = 2;
-            public float TacticsEjectionResistPerPoint = 0;
         }
     }
 }
