+++ Properties/AssemblyInfo.cs
@@ -31,5 +31,4 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 //[assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("3.4.5")]
-[assembly: AssemblyFileVersion("3.4.5")]
+[assembly: AssemblyFileVersion("3.5")]
