+++ assets/mod.json
@@ -3,7 +3,7 @@
 
   "Enabled": true,
 
-  "Version": "3.4.5",
+  "Version": "3.5",
   "Description": "Simulates mechwarrior panic and automatic ejections",
   "Website": "https://github.com/gnivler/PanicSystem",
